consolidate runs models and util functions
