+++ apps/backend/convex/entities/threads/runs/streamAIText.ts
@@ -2,7 +2,7 @@ import { streamText } from 'ai'
 import { internal } from '../../../_generated/api'
 import type { Id } from '../../../_generated/dataModel'
 import type { ActionCtx } from '../../../_generated/server'
-import { hasDelimiter } from '../../../lib/utils'
+import { hasDelimiter } from '../../../lib/parse'
 
 export async function streamAIText(
   ctx: ActionCtx,
