+++ apps/backend/convex/entities/types.ts
@@ -6,7 +6,7 @@ import type { ImageReturn } from './images/validators'
 import type { MessageReturn } from './messages/validators'
 import type { PatternReturn } from './patterns/validators'
 import type { MessageRoles, TextToImageInputs } from './shared'
-import type { RunReturn } from './threads/runs/validators'
+import type { RunReturn } from './threads/runs/models'
 import type { ThreadReturn, ThreadWithDetailsReturn } from './threads/validators'
 import type { UserReturn } from './users/validators'
